new in frame sequences
Gs (in an independent line) have been added (when it has been needed) to have in frame sequences in IgBLAST results with the pattern W/F, G, X, G at the end of the CDR3. 
The second column of the aux table was built based on (0-based) position of the first base of the first complete codon of J sequence. In other words, the reading frame, selecting between:
- reading-frame 0: starting in the first base of J-sequence
- reading-frame 1: starting in the second base of J-sequence
- reading-frame 2: starting in the third base of J-sequence